mcp_sql/requirements.txt
openai
sqlalchemy
psycopg2-binary
fastapi
uvicorn
python-dotenv
asyncpg
alembic
